+++ packages/dashboard/src/components/widgets/AverageResponseTime.tsx
@@ -4,7 +4,7 @@ import { BarChart, Bar, ResponsiveContainer } from "recharts"
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
 import {
     AverageResponseTimeProps
-} from "./getMonitoringData";
+} from "src/app/(layout)/services/[id]/endpoints/[endpointId]/components/monitoring/getMonitoringData";
 import {formatRequestDuration} from "@/utils/formatters";
