+++ packages/dashboard/src/components/widgets/skeletons/skeleton-average-response-time.tsx
