+++ packages/dashboard/src/components/widgets/skeletons/skeleton-cache-hit-ratio.tsx
