+++ packages/dashboard/src/components/widgets/skeletons/skeleton-status-code-distribution.tsx
